hotfix, Paliate embeds and delete //eslint-disable

+++ bot/index.js
@@ -152,7 +152,7 @@ bot.on('message', async message => {
     }
 
     // INSERT INTO DB
-    const sql = 'INSERT INTO stats (content, author_id, author_tag, command, attacker, defender, url, message_id, server_id, will_delete, attacker_description, defender_description, reply_fields, arg) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16)'
+    const sql = 'INSERT INTO stats (content, author_id, author_tag, command, attacker, defender, url, message_id, server_id, will_delete, attacker_description, defender_description, reply_fields, arg) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14)'
     const values = [data.content, data.author_id, data.author_tag, data.command, data.attacker, data.defender, data.url, data.message_id, data.server_id, data.will_delete, data.attacker_description, data.defender_description, data.reply_fields, data.arg]
     await db.query(sql, values)
 
@@ -187,10 +187,10 @@ bot.on('messageReactionAdd', async (reaction, user) => {
   if(returned.rows[0].id === user.id || user.id === meee.id) {
     reaction.message.delete()
       .then().catch(console.error)
-    triggerMessage.delete()
-      .then().catch(console.error)
+    if(triggerMessage)
+      triggerMessage.delete()
+        .then().catch(console.error)
   }
-  triggerMessage
 })
 
 // --------------------------------------
